Added UI microcontroller
renamed the master microcontroller to m for either motor or master or whatever, it just makes things easier for me.

also, the two buttons in the UI code are for transmitting the the code to either stir or toggle the latch.

VERSION 2 should have the two microcontrllers able to communicate together.  probably in the next commit.