VERSION 3
figured out i needed to use 2 power sources per bread board because the motors could not share a source with the LCD and UI C. The stirring re-output on the LCD is fixed, as well as the reinstancing of the update on whether the LCD was saying it was dispensing or not. as of right now, it says dispensing when the button one is pressed, and for 2 seconds after that indefinitely UNLESS the stirring button is pressed.

for the m C, it not default starts turning CCW instead of CW to unturn the bottlecap on my dispensing mechanism.